Refactor and split into separate classes
